refactor: refactorizing security in dlete user.

+++ src/main/java/com/TuWebYa/horno/user/application/usecase/DeleteUserUseCaseImpl.java
@@ -2,6 +2,7 @@ package com.TuWebYa.horno.user.application.usecase;
 
 import com.TuWebYa.horno.user.application.command.DeleteUserCommand;
 import com.TuWebYa.horno.user.application.exception.UserForbiddenException;
+import com.TuWebYa.horno.user.application.exception.UserNotFoundException;
 import com.TuWebYa.horno.user.application.port.in.DeleteUserUseCase;
 import com.TuWebYa.horno.user.application.port.out.UserRepositoryPort;
 import org.springframework.stereotype.Service;
@@ -17,11 +18,21 @@ public class DeleteUserUseCaseImpl implements DeleteUserUseCase {
 
     @Override
     public Mono<Void> deleteUser(DeleteUserCommand command) {
-        if (!command.authenticatedUserId().equals(command.id())
-                && command.authenticatedUserRole().equals("USER")) {
-            return Mono.error(new UserForbiddenException());
-        }
+        return userRepositoryPort.findById(command.id())
+                .switchIfEmpty(Mono.error(new UserNotFoundException("User not found")))
+                .flatMap(userToDelete -> {
+                    if (!command.authenticatedUserId().equals(command.id())
+                            && command.authenticatedUserRole().equals("USER")) {
+                        return Mono.error(new UserForbiddenException());
+                    }
 
-        return userRepositoryPort.deleteById(command.id());
+                    if (command.authenticatedUserRole().equals("ADMIN")
+                            && (userToDelete.getRole() == com.TuWebYa.horno.user.domain.model.UserRole.ADMIN
+                            || userToDelete.getRole() == com.TuWebYa.horno.user.domain.model.UserRole.SUPERADMIN)) {
+                        return Mono.error(new UserForbiddenException("Admin cannot delete another admin or superadmin"));
+                    }
+                    
+                    return userRepositoryPort.deleteById(command.id());
+                });
     }
 }
